requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\SafeTag\requirements.in'
#
asgiref==3.8.1
    # via django
django==5.0.6
    # via
    #   -r .\SafeTag\requirements.in
    #   djangorestframework
django-environ==0.11.2
    # via -r .\SafeTag\requirements.in
djangorestframework==3.15.1
    # via -r .\SafeTag\requirements.in
psycopg2==2.9.9
    # via -r .\SafeTag\requirements.in
sqlparse==0.5.0
    # via django
tzdata==2024.1
    # via django
